:recycle: Refactor features files.
