+++ src/test/java/org/api/testing/demo/steps/booking/common/UnHappyPathsSteps.java
@@ -1,4 +1,4 @@
-package org.api.testing.demo.steps.booking;
+package org.api.testing.demo.steps.booking.common;
 
 import io.cucumber.java.es.Cuando;
 import io.cucumber.java.es.Entonces;
