+++ src/test/resources/features/booking/03_Get_bookings.feature
@@ -11,29 +11,29 @@ Necesidad del negocio: Consultar una reserva
     Antecedentes:
       Dado que la cliente desea consultar la reservación de su próximo viaje
 
-    @getBookingDetails @FunctionalTest
-    Escenario: [YAPE-GB-004] Validar que se pueda consultar una reserva por su BookingId
+    @getBookingDetails @integrationTest
+    Escenario: [YP-004] Validar que se pueda consultar una reserva por su BookingId
       Cuando ella filtre la consulta por el parámetro "ID Booking"
       Entonces visualizará los detalles de la reserva
 
-    @getAllBookings @FunctionalTest
-    Escenario: [YAPE-GB-005] Validar que se puedan consultar todas las reservas existentes
+    @getAllBookings @integrationTest
+    Escenario: [YP-005] Validar que se puedan consultar todas las reservas existentes
       Cuando ella filtre la consulta por el parámetro "All Bookings"
       Entonces visualizará el Id de su reserva
 
-    @getBookingByCustomerNames @FunctionalTest
-    Escenario: [YAPE-GB-006] Validar que se pueda consultar una reserva por los nombres del cliente
+    @getBookingByCustomerNames @integrationTest
+    Escenario: [YP-006] Validar que se pueda consultar una reserva por los nombres del cliente
       Cuando ella filtre la consulta por el parámetro "Customer Names"
       Entonces visualizará el Id de su reserva
 
-    @getBookingByDates @FunctionalTest
-    Escenario: [YAPE-GB-007] Validar que se pueda consultar una reserva por un rango de fechas especificas
+    @getBookingByDates @integrationTest
+    Escenario: [YP-007] Validar que se pueda consultar una reserva por un rango de fechas especificas
       Cuando ella filtre la consulta por el parámetro "Dates"
       Entonces visualizará todos los bookingIDs encontrados para el rango de fechas
       #Entonces visualizará el Id de su reserva (BUG)
 
-    @getBookingBySpecificDates @FunctionalTest
-    Esquema del escenario: [YAPE-GB-008] Ver todos los IDs de las reservas consultando por fechas especificas
+    @getBookingBySpecificDates @integrationTest
+    Esquema del escenario: [YP-008] Ver todos los IDs de las reservas consultando por fechas especificas
       Cuando ella filtre la consulta por las fechas entre "<checkin>" a "<checkout>"
       Entonces visualizará todos los bookingIDs encontrados para el rango de fechas
       Ejemplos:
