+++ src/test/resources/features/booking/05_Delete_booking.feature
@@ -8,8 +8,8 @@ Necesidad del negocio: Eliminar una reserva
 
   Regla de negocio: Para eliminar la información de la reserva, el cliente debe estar autenticado.
 
-    @FunctionalTest
-    Escenario: [YAPE-DL-009] Validar la eliminación de una reserva actual
+    @integrationTest
+    Escenario: [YP-011] Validar la eliminación de una reserva actual
       Dado que la cliente desea eliminar la reservación de su próximo viaje
       Cuando ella solicite la eliminación con su código único de reserva
       Entonces ya no podrá visualizar su reserva en el sistema
