+++ src/test/resources/serenity.conf
@@ -10,7 +10,7 @@ serenity {
 environment = restapi
 
 environments {
-  restapi {
+  dev {
     baseUrl = "https://restful-booker.herokuapp.com"
     auth = "/auth"
     username = "admin"
